Moved tests for cli run control module and renamed/restructured accordingly
