+++ dolosse/acquisition/run_control/cli/examples/control_message.json
