+++ dolosse/acquisition/run_control/cli/examples/manage_message_equipment.json
